feat: extract collect helpers to write-helper
