+++ lib/constants.ts
@@ -1,3 +1,5 @@
+import * as path from 'path'
+
 export const categoryNames: Record<string, string> = {
   api: 'API',
   'api/structures': 'API Structures',
@@ -6,3 +8,5 @@ export const categoryNames: Record<string, string> = {
 }
 
 export const IGNORE_PATTERN = '<!-- i18n-ignore -->'
+
+export const englishBasepath = path.join(__dirname, '..', 'content', 'en-US')
