+++ lib/interfaces/index.ts
@@ -25,3 +25,9 @@ export interface IParseFile extends WalkEntry {
   title?: string
   description?: string
 }
+
+export interface IElectronDocsResponse {
+  slug: string
+  filename: string
+  markdown_content: string
+}
