+++ package.json
@@ -25,6 +25,7 @@
     "del": "^5.0.0",
     "dotenv-safe": "^8.1.0",
     "electron-docs": "^3.0.2",
+    "fs-extra": "^9.0.0",
     "get-crowdin-file-ids": "^1.1.1",
     "github-slugger": "^1.2.0",
     "globals": "^12.0.0",
@@ -39,10 +40,10 @@
     "make-promises-safe": "^5.0.0",
     "match-words": "^1.0.0",
     "mocha": "^7.0.0",
-    "remark-inline-links": "^4.0.0",
-    "remark": "^11.0.2",
     "npm-run-all": "^4.1.2",
     "prettier": "^2.0.0",
+    "remark": "^11.0.2",
+    "remark-inline-links": "^4.0.0",
     "require-yaml": "^0.0.1",
     "rimraf": "^3.0.0",
     "roggy": "^0.1.0",
